enemy's count double to bigint

+++ Assets/Scripts/Enemy.cs
@@ -1,17 +1,18 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Numerics;
 using UnityEngine;
 
 
 [System.Diagnostics.DebuggerDisplay("count: {count}, dice: {dice}")]
 public class Enemy
 {
-    public int count;
+    public BigInteger count;
     public List<int> dice;
     public int id;
 
-    public Enemy(int _count, List<int> _dice, int _id)
+    public Enemy(BigInteger _count, List<int> _dice, int _id)
     {
         count = _count;
         dice = _dice;
